Simplifications and method split for mediaUpload
